Add tests for script options parser
* Extract script options parser into a separate class
* Extract the actual option reslution into a separate method
* Add tests confirming the correct resolution of `{cDefaultRemotePathFromUrl}` script option
