+++ GitUI/GitUI.csproj
@@ -376,6 +376,7 @@
     </Compile>
     <Compile Include="Hotkey\KeysExtensions.cs" />
     <Compile Include="Script\PowerShellHelper.cs" />
+    <Compile Include="Script\ScriptOptionsParser.cs" />
     <Compile Include="Script\ScriptRunner.cs" />
     <Compile Include="AutoCompletion\AutoCompleteWord.cs" />
     <Compile Include="Script\FormFilePrompt.cs">
